+++ src/main/java/com/durhack/sharpshot/registry/entries/StackNodeEntry.kt
@@ -17,7 +17,7 @@ class StackNodeEntry() : RegistryEntry<StackNode>(
                       gc: GraphicsContext,
                       x: Double,
                       y: Double,
-                      scale: Double) {
+                      scale: Int) {
         Draw.triangle(gc, node.direction, x, y, scale, Color.YELLOW)
         Draw.text(gc, "S${node.stackSize}", x, y, scale)
     }
